app/fastapi/requirements_freeze.txt
absl-py==0.8.1
aiofiles==0.4.0
aniso8601==8.0.0
asn1crypto==0.24.0
astor==0.8.0
bcrypt==3.1.7
boto3==1.10.1
botocore==1.13.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.1.4
cvlib==0.2.2
dataclasses==0.7
dnspython==1.16.0
docutils==0.15.2
edgetpu==2.12.1
email-validator==1.0.5
fastapi==0.42.0
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.7
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h11==0.8.1
h5py==2.10.0
httptools==0.0.13
idna==2.6
imageio==2.6.1
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.3
olefile==0.45.1
opencv-python==4.1.1.26
opt-einsum==3.1.0
passlib==1.7.1
Pillow==5.1.0
progressbar==2.5
protobuf==3.10.0
pycparser==2.19
pycrypto==2.6.1
pydantic==0.32.2
pygobject==3.26.1
PyJWT==1.7.1
python-apt==1.6.4
python-dateutil==2.8.0
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.1
SecretStorage==2.3.1
six==1.11.0
starlette==0.12.9
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tflite-runtime==1.14.0
tinydb==3.15.0
typing==3.7.4.1
urllib3==1.25.6
uvicorn==0.9.1
uvloop==0.13.0
websockets==8.0.2
Werkzeug==0.16.0
wrapt==1.11.2
